requirements.txt
anthropic>=0.25.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
streamlit>=1.35.0
